refac(91026): Extrai task gerar_previa_relacao_de_bens_async
